backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
pydantic==2.5.0
python-multipart==0.0.6
groq==0.4.1
httpx==0.25.2
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
slowapi==0.1.9
aiohttp==3.9.1
